Add GeminiProvider; migrating shared logic to LLMProviderBase
