+++ backend/src/dna/llm_providers/llm_provider_base.py
@@ -4,7 +4,10 @@ Abstract base class for LLM providers and factory function.
 """
 
 import os
+from openai import AsyncOpenAI
 from typing import Optional
+
+from dna.prompts.generate_note_prompt import GENERATE_NOTE_PROMPT
 
 
 class LLMProviderBase:
@@ -36,6 +39,42 @@ class LLMProviderBase:
             os.getenv(f"{self.ENV_PREFIX}_TIMEOUT", str(self.DEFAULT_TIMEOUT))
         )
 
+        self._client = None
+
+    @property
+    def client(self) -> AsyncOpenAI:
+        """The interface to the LLM service."""
+        if self._client is None:
+            self._client = self._get_provider_client()
+        return self._client
+
+    def _get_provider_client(self) -> AsyncOpenAI:
+        """Construct an instance of the LLM provider's client."""
+        raise NotImplementedError(f"{self.__class__.__name__} isn't configured.")
+
+    def _substitute_template(
+        self,
+        prompt: str,
+        transcript: str,
+        context: str,
+        existing_notes: str,
+    ) -> str:
+        """Substitute template placeholders in the prompt."""
+        result = prompt
+        result = result.replace("{{ transcript }}", transcript)
+        result = result.replace("{{transcript}}", transcript)
+        result = result.replace("{{ context }}", context)
+        result = result.replace("{{context}}", context)
+        result = result.replace("{{ notes }}", existing_notes)
+        result = result.replace("{{notes}}", existing_notes)
+        return result
+
+    async def close(self) -> None:
+        """Clean up client resources."""
+        if self._client is not None:
+            await self._client.close()
+            self._client = None
+
     async def generate_note(
         self,
         prompt: str,
@@ -56,16 +95,34 @@ class LLMProviderBase:
         Returns:
             The generated note suggestion.
         """
-        raise NotImplementedError()
+        user_message = self._substitute_template(
+            prompt, transcript, context, existing_notes
+        )
 
-    async def close(self) -> None:
-        """Clean up any resources."""
-        pass
+        if additional_instructions:
+            user_message += f"\n\nAdditional Instructions: {additional_instructions}"
+
+        response = await self.client.chat.completions.create(
+            model=self.model,
+            messages=[
+                {"role": "system", "content": GENERATE_NOTE_PROMPT},
+                {"role": "user", "content": user_message},
+            ],
+            temperature=0.7,
+            max_tokens=1024,
+        )
+
+        return response.choices[0].message.content or ""
 
 
 def get_llm_provider() -> LLMProviderBase:
     """Factory function to get the configured LLM provider."""
-    provider_type = os.getenv("LLM_PROVIDER", "openai")
+    provider_type = os.getenv("LLM_PROVIDER", "openai").lower()
+
+    if provider_type == "gemini":
+        from dna.llm_providers.gemini_provider import GeminiProvider
+
+        return GeminiProvider()
 
     if provider_type == "openai":
         from dna.llm_providers.openai_provider import OpenAIProvider
